Ensured No Duplicates
Implemented a system where upon attempted creation of a book, librarian or cardholder, the program checks to see if one of the same name/author etc. already exists. If one already exists, the program does not allow the creation of a duplicate. If one does not already exist, program proceeds with creation.