Some error catching and improvements
Made certain all textboxes have to have information entered in order to attempt creation of book or cardholder or librarian. After the creation of book or cardholder, refreshed combo boxes to now contain those newly added items. Previously you would have to leave, go to a different form and then come back in order to load in the new items. Now upon the button click and successful creation of new book or cardholder, combobox items are refreshed so the new items appear immediately in the drop down list.